refactor: :recycle: Interpreter becomes BabelFish
