+++ src/components/text.js
@@ -1,0 +1,8 @@
+export default function text(data) {
+  return data;
+  /*
+  let newData = data;
+  [...newData.matchAll(/\(\(\((.*)\n(.*)\n\)\)\)/gm)].forEach((match) => {});
+  return newData;
+  */
+}
