+++ src/templates/present/index.js
@@ -1,0 +1,7 @@
+import layoutHTML from "./layout.html.txt";
+import scriptJS from "./script.js.txt";
+import stylesCSS from "./styles.css.txt";
+
+export const view = layoutHTML;
+export const script = scriptJS;
+export const styles = stylesCSS;
