+++ src/utils/toBinary.js
@@ -1,0 +1,9 @@
+const toBinary = (string) => {
+  const codeUnits = new Uint16Array(string.length);
+  for (let i = 0; i < codeUnits.length; i += 1) {
+    codeUnits[i] = string.charCodeAt(i);
+  }
+  return btoa(String.fromCharCode(...new Uint8Array(codeUnits.buffer)));
+};
+
+export default toBinary;
